backend/requirements.txt
django>=4.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
django-cors-headers>=4.0
Pillow>=10.0
python-dotenv>=1.0
psycopg2-binary>=2.9
